+++ src/main/java/com/example/library/dao/categoryDAO/CategoryDAO.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface CategoryDAO {
 
     List<Category> findAll();
-    Category findById(int id);
+    Category findById(long id);
     Category findByName(String name);
     void save(Category category);
-    void deleteById(int id);
+    void deleteById(long id);
 }
