+++ src/main/java/com/example/library/dao/orderDAO/OrderDAO.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface OrderDAO {
     List<Order> findAll();
     void save(Order order);
-    Order findById(int id);
-    List<Order> findByUserId(int userId);
-    Order findLatestByUserId(int userId);
+    Order findById(long id);
+    List<Order> findByUserId(long userId);
+    Order findLatestByUserId(long userId);
     Order findByPayUId(String payUId);
     List<Order> findCompletedOrders();
     List<Order> findCanceledOrders();
