+++ src/main/java/com/example/library/service/categoryService/CategoryService.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public interface CategoryService {
     List<Category> findAll();
-    Category findById(int id);
+    Category findById(long id);
     Category findByName(String name);
     void save(Category category);
-    void deleteById(int id);
+    void deleteById(long id);
 }
