+++ src/main/java/com/example/library/service/categoryService/CategoryServiceImpl.java
@@ -21,7 +21,7 @@ public class CategoryServiceImpl implements CategoryService{
     }
 
     @Override
-    public Category findById(int id) {
+    public Category findById(long id) {
         return categoryDAO.findById(id);
     }
 
@@ -36,7 +36,7 @@ public class CategoryServiceImpl implements CategoryService{
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(long id) {
         categoryDAO.deleteById(id);
     }
 }
